requirements.txt
streamlit
google-generativeai
streamlit_option_menu
streamlit_extras